Settings: Fix battery usage chart irregular
* getBatteryLevel passed a value of 0 or even lower,
  which is inconsistent with the actual value, which
  will cause the chart to be irregular.

Test: Clear SettingsIntelligence, wait 2 hours

+++ src/com/google/android/settings/fuelgauge/DatabaseUtils.java
@@ -194,7 +194,7 @@ public final class DatabaseUtils {
         if (scale == 0) {
             return -1;
         }
-        return Math.round((level / scale) * 100.0f);
+        return Math.round((float) level / (float) scale * 100.0f);
     }
 
     private static SharedPreferences getSharedPreferences(Context context) {
